create structs "attribution", "OrderRecipe", create enums "RecipeNames", "RecipeTypes", updates on some models. Clients can alr command something. Implement thread pool and solve concurrency issues in DinningRoomController.h

+++ src/Models/dinningRoomModels/classDeclaration/headWaiter.h
@@ -17,7 +17,7 @@ class HeadWaiter : public MobileElementModel {
     queue<Order> orderQueue;
 public:
 
-    HeadWaiter(int abscice, int intercept);
+    HeadWaiter(double abscice, double intercept);
     /**
      * @brief the function to lead the clients to their table
      *
